Added AccelerometerSensor and SmartphoneDevice
Fixed some types and improoved some names
Removed Booleaned class
Added generics and timestamp to the sensor event entry

+++ app/src/androidTest/java/eu/fbk/mpba/sensorsflows/stubs/SensorStub.java
@@ -18,7 +18,7 @@ import eu.fbk.mpba.sensorsflows.base.SensorStatus;
  *      public void sensorEvent(TimeT time, int type, String message)
  * protetti quindi visibili nella sottoclasse.
  */
-public class SensorStub extends SensorImpl<Integer, Double[]> {
+public class SensorStub extends SensorImpl<Long, float[]> {
 
     /**
      * Costruttore pienamente personalizzato
